feat: Raise minimum supported versions
BREAKING CHANGE: Node 10 is EOL next month, so I'm taking the
opportunity to batch together some breaking changes / end of
support.

- Raise the minimum Node version to 12.10
- Drop the Brotli plugin in favour of Nodes native support
- Update the package.json "files" list to make the install a little
  smaller
- Update the `README` to drop the docs around earlier versions of Nuxt;
  I don't want to attempt to guarantee to support for them

For most use cases this update is probably a no-op, unless you're on a
very old version of Node or Nuxt. If you were overriding the default
options for the Brotli compression, you may need to change your config
slightly; see
https://webpack.js.org/plugins/compression-webpack-plugin/#using-brotli
for an example of the options that can be passed through.
